Merge TReader and TWriter
Now both read and write are done in the same structure. It works for
files and buffers. The stream and its operation are customizable. The
tokenizer available in treader has been removed since it should be
a separate entity.

TError has been improved a little so that TRW can report issues.